[refactor] Convention 모듈 내 헬퍼 함수들 상위 패키지로 이동
