+++ build-logic/convention/src/main/java/com/ssavice/AndroidCompose.kt
@@ -1,4 +1,4 @@
-package com.ssavice.convention
+package com.ssavice
 
 import com.android.build.api.dsl.CommonExtension
 import org.gradle.api.Project
