+++ build-logic/convention/src/main/java/com/ssavice/ProjectExtensions.kt
@@ -1,4 +1,4 @@
-package com.ssavice.convention
+package com.ssavice
 
 import org.gradle.api.Project
 import org.gradle.api.artifacts.VersionCatalog
