+++ build-logic/convention/src/main/kotlin/AndroidApplicationComposeConventionPlugin.kt
@@ -3,7 +3,7 @@ import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
 import org.gradle.kotlin.dsl.getByType
 import com.android.build.api.dsl.ApplicationExtension
-import com.ssavice.convention.configureAndroidCompose
+import com.ssavice.configureAndroidCompose
 
 class AndroidApplicationComposeConventionPlugin : Plugin<Project> {
     override fun apply(target: Project) {
