+++ build-logic/convention/src/main/kotlin/AndroidApplicationConventionPlugin.kt
@@ -3,7 +3,7 @@ import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
 import org.gradle.kotlin.dsl.getByType
-import com.ssavice.convention.configureKotlinAndroid
+import com.ssavice.configureKotlinAndroid
 
 class AndroidApplicationConventionPlugin : Plugin<Project> {
     override fun apply(target: Project) {
